refactor(core): migrate core interfaces and refactor project modules
- Move global core interfaces (ID, DateTimeString, EntityAttribute, EntityStatus) to the design-system package
- Remove shared-core project store and interfaces, updating apps to consume centralized design-system core types
- Update router paths and component imports for project views across apps/auth and apps/mull
