--- apps/auth/src/components/organisms/login/LoginModal.vue
@@ -1,118 +1,0 @@
-<template>
-  <VModal :is-open="isOpen" size="md" @close="emit('close')">
-    <VBox tag="header" intent="brand" surface="solid" padding="lg" border="bottom" class="text-center">
-      <VStack align="center" gap="md">
-        <VBox padding="sm" intent="brand" surface="solid" rounded="2xl" border="all" :inline="true" class="shadow-sm">
-          <VIcon name="LockClosed" size="xs" />
-        </VBox>
-
-        <VStack gap="xs">
-          <VTypography tag="h2" size="2xl" weight="bold">
-            User Login
-          </VTypography>
-          <VTypography size="sm">
-            Access your research initiatives and strategic workspace.
-          </VTypography>
-        </VStack>
-      </VStack>
-    </VBox>
-
-    <VBox padding="lg">
-      <VForm gap="md" @submit="handleLogin">
-
-        <VFormField label="Username" v-slot="{ id }">
-          <VInput :id="id" v-model="username" :disabled="isSubmitting" required />
-        </VFormField>
-
-        <VFormField label="Password" v-slot="{ id }">
-          <VInput :id="id" v-model="password" type="password" :disabled="isSubmitting" required />
-        </VFormField>
-
-        <transition name="fade">
-          <VAlert v-if="error" intent="danger" surface="soft">
-            <VCluster gap="xs">
-              <VIcon name="ExclamationCircle" size="sm" />
-              <VTypography size="xs" intent="danger" surface="soft">{{ error }}</VTypography>
-            </VCluster>
-          </VAlert>
-        </transition>
-
-        <VBox class="pt-2">
-          <VButton
-            type="submit"
-            attention="primary"
-            size="lg"
-            class="w-full"
-            :disabled="isSubmitting"
-          >
-            Sign In
-          </VButton>
-        </VBox>
-
-      </VForm>
-    </VBox>
-
-    <VBox tag="footer" padding="lg" class="pt-0 text-center">
-      <VTypography size="xs" attention="secondary">
-        By logging in, you agree to our research workspace protocols.
-      </VTypography>
-    </VBox>
-  </VModal>
-</template>
-
-<script setup lang="ts">
-import { useAuthStore } from '@auraflux/shared-core/stores/auth';
-import { ref } from 'vue';
-import { useRouter } from 'vue-router';
-
-// Atoms & Molecules
-import VButton from '@auraflux/design-system/components/atoms/buttons/VButton.vue';
-import VInput from '@auraflux/design-system/components/atoms/forms/VInput.vue';
-import VIcon from '@auraflux/design-system/components/atoms/indicators/VIcon.vue';
-import VTypography from '@auraflux/design-system/components/atoms/indicators/VTypography.vue';
-import VBox from '@auraflux/design-system/components/atoms/layout/VBox.vue';
-import VCluster from '@auraflux/design-system/components/atoms/layout/VCluster.vue';
-import VStack from '@auraflux/design-system/components/atoms/layout/VStack.vue';
-import VForm from '@auraflux/design-system/components/molecules/forms/VForm.vue';
-import VFormField from '@auraflux/design-system/components/molecules/forms/VFormField.vue';
-import VAlert from '@auraflux/design-system/components/molecules/indicators/VAlert.vue';
-import VModal from '@auraflux/design-system/components/molecules/indicators/VModal.vue';
-
-const props = defineProps<{
-  isOpen: boolean;
-}>();
-
-const emit = defineEmits<{
-  (e: 'close'): void;
-}>();
-
-const router = useRouter();
-const authStore = useAuthStore();
-
-const username = ref('');
-const password = ref('');
-const error = ref('');
-const isSubmitting = ref(false);
-
-const handleLogin = async () => {
-  if (isSubmitting.value) return;
-  error.value = '';
-  isSubmitting.value = true;
-
-  try {
-    const success = await authStore.loginUser(username.value, password.value);
-    if (success) {
-      username.value = '';
-      password.value = '';
-      emit('close');
-      router.push('/projects/')
-    } else {
-      error.value = 'Invalid username or password.';
-    }
-  } catch (err: any) {
-    error.value = err.message || 'Authentication server unreachable.';
-  } finally {
-    isSubmitting.value = false;
-  }
-};
-</script>
