+++ apps/mull/src/router/index.ts
@@ -1,5 +1,6 @@
 import LandingPage from '@/components/pages/LandingPage.vue';
 import MainPage from '@/components/pages/MainPage.vue';
+import ProjectPage from '@/components/pages/ProjectPage.vue';
 import RepositoryPage from '@/components/pages/RepositoryPage.vue';
 import RoutePage from '@/components/pages/RoutePage.vue';
 import { createRouter, createWebHistory, RouteRecordRaw } from 'vue-router';
@@ -11,17 +12,22 @@ export const routes: Array<RouteRecordRaw> = [
     component: LandingPage
   },
   {
-    path: '/projects/:id/mull',
+    path: '/projects/',
     name: 'RoutePage',
     component: RoutePage,
     children: [
       {
         path: '',
+        name: 'ProjectPage',
+        component: ProjectPage
+      },
+      {
+        path: ':id',
         name: 'MainPage',
         component: MainPage
       },
       {
-        path: 'repository',
+        path: ':id/repository',
         name: 'RepositoryPage',
         component: RepositoryPage
       }
